third_party/WebKit/LayoutTests/css3/motion-path/use-count-offset.html

<!DOCTYPE html>
<head>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
  <style>
  .n { offset: none; }
  </style>
</head>

<div id="target"></div>
<script>
var CSSOffsetInEffect = 1617; // From UseCounter.h

test(function() {
  assert_false(internals.isUseCounted(document, CSSOffsetInEffect));

  var styleElement = document.createElement('style');
  styleElement.textContent = ".p { offset: path('m 0 0 h 1') 0px auto; }";
  document.head.appendChild(styleElement);

  assert_true(internals.isUseCounted(document, CSSOffsetInEffect));
});
</script>
